backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
Pillow
rasterio
torch
torchvision
segmentation-models-pytorch
